+++ lib/main.dart
@@ -1,26 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:training_app/ui/sign_in_page/sign_in_page.dart';
+import 'package:training_app/ui/root_page/root_page.dart';
 
 void main() {
-  runApp(const MyApp());
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      debugShowCheckedModeBanner: false,
-      title: 'Dear Diary',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const Scaffold(
-          backgroundColor: Colors.blue,
-          body: Center(
-            child: SignInPage(),
-          )),
-    );
-  }
+  runApp(const DearDiaryApp());
 }
