+++ lib/ui/root_page/root_page.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:training_app/ui/sign_in_page/sign_in_page.dart';
+
+class DearDiaryApp extends StatelessWidget {
+  const DearDiaryApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      debugShowCheckedModeBanner: false,
+      title: 'Dear Diary',
+      theme: ThemeData(
+        primarySwatch: Colors.blue,
+      ),
+      home: const Scaffold(
+          backgroundColor: Colors.blue,
+          body: Center(
+            child: SignInPage(),
+          )),
+    );
+  }
+}
